+++ src/main/java/racingcar/controller/CarRacingGameController.java
@@ -4,6 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 import racingcar.model.Car;
 import racingcar.model.WinnerDecision;
+import racingcar.util.InputValidator;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
