+++ src/main/java/racingcar/util/InputValidator.java
@@ -1,4 +1,4 @@
-package racingcar.controller;
+package racingcar.util;
 
 public class InputValidator {
